ISSUE-25 fix names and move testhelpers
