+++ core/users/api/e2e/e2e.go
@@ -121,16 +121,16 @@ func newTestRepository(t *testing.T, cfg Cfg) (*database.Repository, func()) {
 		t.Fatal(err)
 	}
 
-	if err := schema.Seed(repo.Storer, "users"); err != nil {
+	if err := schema.Seed(repo.SqlxStorer, "users"); err != nil {
 		t.Fatal(err)
 	}
-	if err := schema.Seed(repo.Storer, "teams"); err != nil {
+	if err := schema.Seed(repo.SqlxStorer, "teams"); err != nil {
 		t.Fatal(err)
 	}
-	if err := schema.Seed(repo.Storer, "memberships"); err != nil {
+	if err := schema.Seed(repo.SqlxStorer, "memberships"); err != nil {
 		t.Fatal(err)
 	}
-	if err := schema.Seed(repo.Storer, "projects"); err != nil {
+	if err := schema.Seed(repo.SqlxStorer, "projects"); err != nil {
 		t.Fatal(err)
 	}
 	return repo, rClose
