+++ core/users/platform/testhelpers/testhelpers.go
@@ -1,4 +1,4 @@
-package handlers
+package testhelpers
 
 func StringPointer(s string) *string {
 	return &s
